style: workspace clippy autofix sweep
Bulk run of `cargo clippy --fix --workspace --tests` to clear mechanical
lints across the parity binaries, runtime input modules, and several
machine cores. 47 files, all autofix-only — no behaviour change, no
test edits.

The fixed lints, sorted by frequency:

- collapsible_if (21 sites) — let-chain / match-guard rewrites across
  every per-machine binary's MCP + script default-path resolvers.
- needless_range_loop (4 sites) — indexed `for i in 0..N` loops over
  fixed-size arrays in input-event maps and chip-tick paths.
- manual_is_multiple_of (4 sites) — `% 2 == 0` style checks in chip
  divider logic now use `is_multiple_of`.
- unnecessary_cast (2 sites) — same-type `u16` / `u32` casts in
  mos-vic-i and sega-vdp.
- collapsible_match (1), unnecessary_min_or_max (1),
  unnecessary_map_or (1), manual_range_patterns (1),
  manual_checked_ops (1), doc_lazy_continuation (1),
  derivable_impls (1).

Verified: cargo build clean across the touched crates, targeted
`cargo clippy --no-deps -- -D warnings` on each is green.

Remaining workspace clippy issues (unwrap_used in test code, several
warnings in ti-tms9918 / ti-sn76489 / etc.) are deferred to a follow-up.

+++ crates/mos-vic-i/src/lib.rs
@@ -122,15 +122,15 @@ impl Vic6560 {
             let char_col = self.pixel_x;
 
             // Screen memory base from registers
-            let screen_base = ((u16::from(self.regs[5]) & 0xF0) as u16) << 6
-                | ((u16::from(self.regs[2]) & 0x80) as u16) << 2;
+            let screen_base = (u16::from(self.regs[5]) & 0xF0) << 6
+                | (u16::from(self.regs[2]) & 0x80) << 2;
 
             let char_addr = screen_base.wrapping_add(char_row as u16 * 22 + char_col as u16);
             let char_code = read_screen(char_addr);
             let colour_nibble = read_colour(char_addr) & 0x0F;
 
             // Character ROM lookup
-            let char_rom_base = ((u16::from(self.regs[5]) & 0x0F) as u16) << 10;
+            let char_rom_base = (u16::from(self.regs[5]) & 0x0F) << 10;
             let char_rom_addr = char_rom_base.wrapping_add(u16::from(char_code) * 8 + pixel_in_char_y as u16);
             let char_data = read_char_rom(char_rom_addr);
